requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
langchain-ollama>=0.2.0
sib-api-v3-sdk>=7.6.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
email-validator>=2.0.0
python-dotenv>=1.0.1
requests>=2.32.0
httpx>=0.27.0
numpy>=2.0.0
imapclient>=3.0.0
apscheduler>=3.10.0

# Microsoft Graph API
msal>=1.24.0

# Database
sqlalchemy>=2.0.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9

# UI
streamlit>=1.30.0
